Move helpers to Aspite.Hosting.Azure
