+++ src/Aspire.Hosting.Azure/Aspire.Hosting.Azure.csproj
@@ -14,7 +14,6 @@
     <Compile Include="..\Aspire.Hosting\Dcp\Process\ProcessSpec.cs" Link="Provisioning\Utils\ProcessSpec.cs" />
     <Compile Include="..\Aspire.Hosting\Dcp\Process\ProcessUtil.cs" Link="Provisioning\Utils\ProcessUtil.cs" />
     <Compile Include="..\Shared\Cosmos\CosmosConstants.cs" Link="Shared\CosmosConstants.cs" />
-    <Compile Include="$(SharedDir)ResourceGroupNameHelpers.cs" Link="shared\ResourceGroupNameHelpers.cs" />
   </ItemGroup>
 
   <ItemGroup>
