+++ src/Aspire.Hosting.Azure/Provisioning/Provisioners/AzureProvisioner.cs
@@ -7,8 +7,8 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using Aspire.Hosting.ApplicationModel;
 using Aspire.Hosting.Azure.Provisioning;
+using Aspire.Hosting.Azure.Utils;
 using Aspire.Hosting.Lifecycle;
-using Aspire.Hosting.Utils;
 using Azure;
 using Azure.Core;
 using Azure.Identity;
